moved file locations and updated component testing for icons
