+++ src/features/CourtCaseList/NotePreviewButton.tsx
@@ -40,26 +40,24 @@ export const NotePreview = ({ latestNote, displayDate, numberOfNotes }: NotePrev
   )
 }
 
-const NotePreviewButton: React.FC<NotePreviewButtonProps> = (props: NotePreviewButtonProps) => {
+export const PreviewButton = () => {
+  return (
+    <>
+      <span className="govuk-accordion-nav__chevron govuk-accordion-nav__chevron--down"></span>
+      <span className="govuk-accordion__show-all-text">{"Preview"}</span>
+    </>
+  )
+}
+export const HideButton = () => {
+  return (
+    <>
+      <span className="govuk-accordion-nav__chevron"></span>
+      <span className="govuk-accordion__show-all-text">{"Hide"}</span>
+    </>
+  )
+}
+export const NotePreviewButton: React.FC<NotePreviewButtonProps> = (props: NotePreviewButtonProps) => {
   const classes = useStyles()
-
-  const PreviewButton = () => {
-    return (
-      <>
-        <span className="govuk-accordion-nav__chevron govuk-accordion-nav__chevron--down"></span>
-        <span className="govuk-accordion__show-all-text">{"Preview"}</span>
-      </>
-    )
-  }
-
-  const HideButton = () => {
-    return (
-      <>
-        <span className="govuk-accordion-nav__chevron"></span>
-        <span className="govuk-accordion__show-all-text">{"Hide"}</span>
-      </>
-    )
-  }
 
   return (
     <>
@@ -79,5 +77,3 @@ const NotePreviewButton: React.FC<NotePreviewButtonProps> = (props: NotePreviewB
     </>
   )
 }
-
-export default NotePreviewButton
